pyproject.toml
[tool.poetry]
name = "simian"
version = "0.1.0"
description = ""
authors = ["foxgem <jianhgreat@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic-ai = {extras = ["logfire"], version = "^0.0.12"}
python-dotenv = "^1.0.1"
pydantic-ai-slim = {extras = ["groq"], version = "^0.0.12"}
duckduckgo-search = "^6.4.1"
markdown-pdf = "^1.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
